+++ incapsula/resource_acl_security_rule.go
@@ -128,10 +128,19 @@ func resourceACLSecurityRuleRead(d *schema.ResourceData, m interface{}) error {
 
 	siteStatusResponse, err := client.SiteStatus("acl-rule-read", d.Get("site_id").(int))
 
+	// Site object may have been deleted
+	if siteStatusResponse != nil && siteStatusResponse.Res.(float64) == 9413 {
+		log.Printf("[INFO] Incapsula Site ID %s has already been deleted: %s\n", d.Get("site_id"), err)
+		d.SetId("")
+		return nil
+	}
+
 	if err != nil {
 		log.Printf("[ERROR] Could not read Incapsula ACL Rule for id: %s, %s\n", ruleID, err)
 		return err
 	}
+
+	found := false
 
 	// Now with the site status, iterate through the rules and find our ID
 	for _, entry := range siteStatusResponse.Security.Acls.Rules {
@@ -155,8 +164,15 @@ func resourceACLSecurityRuleRead(d *schema.ResourceData, m interface{}) error {
 			case whitelistedIPs:
 				d.Set("ips", strings.Join(entry.Ips, ","))
 			}
+			found = true
 			break
 		}
+	}
+
+	if !found {
+		log.Printf("[INFO] Incapsula ACL Security Rule ID %s for Site ID %d has already been deleted: %s\n", ruleID, d.Get("site_id").(int), err)
+		d.SetId("")
+		return nil
 	}
 
 	log.Printf("[INFO] Read Incapsula ACL Rule for id: %s\n", ruleID)